+++ voxidian-editor-common/src/packet/c2s/open_file.rs
@@ -3,7 +3,7 @@ use super::*;
 
 #[derive(Debug)]
 pub struct OpenFileC2SPacket {
-    pub id : u32
+    pub id : u64
 }
 
 impl PacketMeta for OpenFileC2SPacket {
